Commit Day 1 Java Core

+++ Java Assignment/Lesson_1/src/DefaultPackage/Account.java
@@ -3,7 +3,7 @@ package DefaultPackage;
 import java.util.Date;
 
 public class Account {
-    byte id;
+    int id;
     String email;
     String userName;
     String fullName;
